requirements.txt
aniso8601==3.0.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
gunicorn==19.8.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.1
pymongo==3.6.1
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
